+++ src/types/custom-attributes.d.ts
@@ -11,26 +11,8 @@
  */
 
 
-/* THIS FILE IS AUTO-GENERATED - SEE CONTRIBUTOR DOCUMENTATION */
-
-import { Resource } from '../resource';
-import { CustomAttributeValue } from '../custom-attributes';
-import { Client } from '../client';
-import { OptionalKnownProperties } from '../optional-known-properties-type';
-
-
-declare class GroupProfile extends Resource {
-  constructor(resourceJson: Record<string, unknown>, client: Client);
-
-  description: string;
-  name: string;
-  [key: string]: CustomAttributeValue | CustomAttributeValue[]
-
-}
-
-type GroupProfileOptions = OptionalKnownProperties<GroupProfile>;
+type CustomAttributeValue = boolean | number | string;
 
 export {
-  GroupProfile,
-  GroupProfileOptions
+  CustomAttributeValue
 };