+++ CHANGELOG.md
@@ -4,8 +4,8 @@
 
 ### Bug Fixes
 
-- [#329](https://github.com/okta/okta-sdk-nodejs/pull/360) Fixes path and type signature issues for MemoryStore.
-
+- [#360](https://github.com/okta/okta-sdk-nodejs/pull/360) Fixes path and type signature issues for MemoryStore.
+- [#365](https://github.com/okta/okta-sdk-nodejs/pull/365) Adds support for custom attributes in UserProfile and GroupProfile
 
 # 6.5.0